backend/rag_service/requirements.txt
flask
google-generativeai
faiss-cpu
numpy
python-dotenv
requests 